backend/requirements.txt
flask==3.0.0
google-cloud-storage==2.13.0
google-cloud-firestore==2.14.0
